backend/requirements_colorize.txt
fastapi
uvicorn
python-dotenv
numpy
Pillow
torch
torchvision
scikit-image
scikit-learn
